move `RegionResolutionError` into `lexical_region_resolve`
